fix: move bonus code inside bonus folder
