+++ src/bonus/components/BonusPriceTag.tsx
