+++ src/bonus/components/index.tsx
@@ -1,0 +1,2 @@
+export {PayWithBonusPointsCheckbox} from './PayWithBonusPointsCheckbox';
+export {BonusPriceTag} from './BonusPriceTag';
