+++ src/bonus/index.tsx
@@ -1,1 +1,5 @@
-export {BonusPriceTag} from './BonusPriceTag';
+export {BonusPriceTag, PayWithBonusPointsCheckbox} from './components';
+
+export {isActive, findRelevantBonusProduct} from './utils';
+
+export {getBonusBalanceQuery} from './queries/get-bonus-balance-query';
